+++ common/src/main/java/net/mehvahdjukaar/polytone/IrisCompat.java
@@ -1,0 +1,18 @@
+package net.mehvahdjukaar.polytone;
+
+import net.irisshaders.iris.Iris;
+import net.irisshaders.iris.pipeline.ShaderRenderingPipeline;
+import net.irisshaders.iris.pipeline.WorldRenderingPipeline;
+
+public class IrisCompat {
+    public IrisCompat() {
+    }
+
+    public static boolean isIrisShaderFuckerActive() {
+        WorldRenderingPipeline pipeline = Iris.getPipelineManager().getPipelineNullable();
+        if (pipeline instanceof ShaderRenderingPipeline s) {
+            return s.shouldOverrideShaders();
+        }
+        return false;
+    }
+}